+++ muna/beta/openai/annotations.py
@@ -68,10 +68,140 @@ class Annotations:
         )
 
     @classmethod
+    def FrequencyPenalty(
+        cls,
+        *,
+        description: str,
+        min: float | None=None,
+        max: float | None=None,
+        **kwargs
+    ) -> Parameter:
+        """
+        Frequency penalty parameter.
+        """
+        return Parameter(
+            name="",
+            description=description,
+            denotation="openai.chat.completions.frequency_penalty",
+            range=(min, max) if min is not None and max is not None else None,
+            **kwargs
+        )
+
+    @classmethod
+    def MaxOutputTokens(
+        cls,
+        *,
+        description: str,
+        **kwargs
+    ) -> Parameter:
+        """
+        Maximum output tokens parameter.
+        """
+        return Parameter(
+            name="",
+            description=description,
+            denotation="openai.chat.completions.max_output_tokens",
+            **kwargs
+        )
+
+    @classmethod
+    def PresencePenalty(
+        cls,
+        *,
+        description: str,
+        min: float | None=None,
+        max: float | None=None,
+        **kwargs
+    ) -> Parameter:
+        """
+        Presence penalty parameter.
+        """
+        return Parameter(
+            name="",
+            description=description,
+            denotation="openai.chat.completions.presence_penalty",
+            range=(min, max) if min is not None and max is not None else None,
+            **kwargs
+        )
+
+    @classmethod
+    def ReasoningEffort(
+        cls,
+        *,
+        description: str,
+        **kwargs
+    ) -> Parameter:
+        """
+        Reasoning effort parameter.
+        """
+        return Parameter(
+            name="",
+            description=description,
+            denotation="openai.chat.completions.reasoning_effort",
+            **kwargs
+        )
+    
+    @classmethod
+    def ResponseFormat(
+        cls,
+        *,
+        description: str,
+        **kwargs
+    ) -> Parameter:
+        """
+        Response format parameter.
+        """
+        return Parameter(
+            name="",
+            description=description,
+            denotation="openai.chat.completions.response_format",
+            **kwargs
+        )
+
+    @classmethod
+    def ResponseStreaming(
+        cls,
+        *,
+        description: str,
+        **kwargs
+    ) -> Parameter:
+        """
+        Response streaming parameter.
+        """
+        return Parameter(
+            name="",
+            description=description,
+            denotation="openai.chat.completions.stream",
+            **kwargs
+        )
+
+    @classmethod
+    def SamplingProbability(
+        cls,
+        *,
+        description: str,
+        min: float | None=None,
+        max: float | None=None,
+        **kwargs
+    ) -> Parameter:
+        """
+        Sampling probability parameter.
+        """
+        return Parameter(
+            name="",
+            description=description,
+            denotation="openai.chat.completions.top_p",
+            range=(min, max) if min is not None and max is not None else None,
+            **kwargs
+        )
+
+    @classmethod
     def SamplingTemperature(
         cls,
         *,
         description: str,
+        min: float | None=None,
+        max: float | None=None,
         **kwargs
     ) -> Parameter:
         """
@@ -81,6 +211,7 @@ class Annotations:
             name="",
             description=description,
             denotation="openai.chat.completions.temperature",
+            range=(min, max) if min is not None and max is not None else None,
             **kwargs
         )